feat: complete exercise 3
use NextJS static metadata
object and dynamic
metadata function to
populate page metadata

update `loadBlogPost` to
be wrapped in `React.cache`. Initially,
I did this within the `page.js`
component because I
didn't want to import React
to the helper file. i.e. import
`loadBlogPost` in `page.js`
and create a `cachedLoadBlogPost` fn
that wraps the imported fn
with `React.cache`.

Ultimately, I updated it to match
the solution because it means we don't have to wrap it
everywhere else we use `loadBlogPost`