feat: implement exercise 5C
This exercise required implementing
a way to ensure the index was consistent
between the division group areas and
the remainder area.

Framer will continue to animate things smoothly
as long as each index is sequential within the
LayoutGroup.

We use reverse to ensure the items are added
at the end of the remainder area.